Create TASK1_NUMBER GUESS
The code you provided is a number guessing game implemented in Java. Here's a brief overview of its functionality:

The program starts by displaying a welcome message to the user.
It enters a while loop that allows the user to play multiple rounds of the game.
Within each round, a random number between 1 and 100 is generated using the Random class.
The user is prompted to guess the generated number.
The program provides feedback to the user, indicating whether the guess is too low or too high.
The user has a maximum of 10 attempts to guess the correct number.
If the user guesses the correct number within the given attempts, a congratulatory message is displayed along with the number of attempts it took.
If the user runs out of attempts without guessing the correct number, a message is displayed with the correct number.
After each round, the user is asked if they want to play again. If the user responds with "yes," another round starts. If the response is "no," the game ends.
Finally, a closing message is displayed to thank the user for playing the game.
The code combines the use of the Scanner class to read user input, the Random class to generate random numbers, and control flow structures such as loops and conditionals to implement the game logic.